+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 * Remove the dependency on jQuery
 * Add a validate-element-location rule that warns when elements appear as descendants of elements they're not allowed to descend from.
-* Add the ability to exclude DOM elements and subtrees from inspection.
+* Add the ability to exclude DOM elements and DOM subtrees from inspection via the `exclude` and `excludeSubTree` config options.
 
 == 0.2.3 (June 15, 2013)